+++ src/main/java/com/growup/pms/invitation/controller/dto/request/TeamInvitationCreateRequest.java
@@ -1,6 +1,6 @@
 package com.growup.pms.invitation.controller.dto.request;
 
-import com.growup.pms.invitation.domian.dto.TeamInvitationCreateCommand;
+import com.growup.pms.invitation.service.dto.TeamInvitationCreateCommand;
 import jakarta.validation.constraints.NotNull;
 
 public record TeamInvitationCreateRequest(@NotNull Long userId) {
