+++ src/main/java/com/growup/pms/invitation/service/TeamInvitationService.java
@@ -4,8 +4,8 @@ import com.growup.pms.common.exception.code.ErrorCode;
 import com.growup.pms.common.exception.exceptions.DuplicateException;
 import com.growup.pms.common.exception.exceptions.EntityNotFoundException;
 import com.growup.pms.invitation.domian.TeamInvitation;
-import com.growup.pms.invitation.domian.dto.TeamInvitationCreateCommand;
 import com.growup.pms.invitation.repository.TeamInvitationRepository;
+import com.growup.pms.invitation.service.dto.TeamInvitationCreateCommand;
 import com.growup.pms.role.domain.TeamRole;
 import com.growup.pms.role.repository.RoleRepository;
 import com.growup.pms.team.domain.TeamUser;
