+++ src/test/java/com/growup/pms/invitation/service/TeamInvitationServiceTest.java
@@ -11,8 +11,8 @@ import static org.mockito.Mockito.when;
 import com.growup.pms.common.exception.exceptions.DuplicateException;
 import com.growup.pms.common.exception.exceptions.EntityNotFoundException;
 import com.growup.pms.invitation.domian.TeamInvitation;
-import com.growup.pms.invitation.domian.dto.TeamInvitationCreateCommand;
 import com.growup.pms.invitation.repository.TeamInvitationRepository;
+import com.growup.pms.invitation.service.dto.TeamInvitationCreateCommand;
 import com.growup.pms.role.domain.Role;
 import com.growup.pms.role.domain.RoleType;
 import com.growup.pms.role.domain.TeamRole;
